Replace add_user and dev_jwt_generator scripts
Removed add_user.py and dev_jwt_generator.py, and added add_user_standalone.py with improved CLI, environment loading, and direct database access. Updated refresh_tokens.user_id type in init_db.sql from INT to BIGINT for consistency with user_profiles.

+++ bot/database/init_db.sql
@@ -270,7 +270,7 @@ END $$;
 
 CREATE TABLE IF NOT EXISTS refresh_tokens (
     id SERIAL PRIMARY KEY,
-    user_id INT REFERENCES user_profiles(user_id) ON DELETE CASCADE,
+    user_id BIGINT REFERENCES user_profiles(user_id) ON DELETE CASCADE,
     token VARCHAR(255) UNIQUE NOT NULL,
     created_at TIMESTAMPTZ DEFAULT NOW(),
     expires_at TIMESTAMPTZ NOT NULL,
